+++ src/index.ts
@@ -4,6 +4,8 @@ import EventEmitter from 'eventemitter3';
 import assert from 'assert';
 import { createRequire } from 'node:module';
 
+export * from './types/index';
+
 if (typeof process !== 'undefined' && parseInt(process.versions.node.split('.')[0]) < 16) {
   console.error('Your node version is currently', process.versions.node);
   console.error('Please update it to a version >= 16.x.x from https://nodejs.org/');
@@ -119,38 +121,3 @@ export class Swarm extends EventEmitter {
     return Object.keys(this.plugins).includes(name);
   }
 }
-
-/**
- * @callback Plugin A plugin that can be loaded into a bot.
- * @param {Bot} The bot to load the plugin into.
- * @param {ClientOptions} [opts] The bot's ClientOptions.
- */
-type Plugin = (((bot: Bot) => null) | ((bot: Bot, opts: ClientOptions) => null));
-
-/**
- * @typedef {Object} BotSwarmData - Data about bots stored by the swarm.
- * @property {ClientOptions} [botOptions] - The bot's ClientOptions.
- * @property {boolean} injectAllowed - Whether the bot is ready for plugin injection.
- */
-export class BotSwarmData {
-  botOptions?: ClientOptions;
-  injectAllowed = false;
-}
-
-/**
- * @typedef {Object} SwarmBot - A bot in a swarm.
- * @property {BotSwarmData} [swarmOptions] - The bot's BotSwarmData.
- */
-export interface SwarmBot extends Bot {
-  swarmOptions?: BotSwarmData
-}
-
-/**
- * @typedef {Partial<ClientOptions>} AuthenticationOptions - Authentication options for swarms.
- */
-export type AuthenticationOptions = Partial<ClientOptions>;
-
-/**
- * @typedef {Partial<ClientOptions>} ConnectionOptions - Connection options for swarms.
- */
-export type ConnectionOptions = Partial<ClientOptions>;
