+++ src/types/authenticationOptions.ts
@@ -1,0 +1,6 @@
+import { ClientOptions } from 'minecraft-protocol';
+
+/**
+ * @typedef {Partial<ClientOptions>} AuthenticationOptions - Authentication options for swarms.
+ */
+export type AuthenticationOptions = Partial<ClientOptions>;
