+++ src/types/connectionOptions.ts
@@ -1,0 +1,6 @@
+import { ClientOptions } from 'minecraft-protocol';
+
+/**
+ * @typedef {Partial<ClientOptions>} ConnectionOptions - Connection options for swarms.
+ */
+export type ConnectionOptions = Partial<ClientOptions>;
