+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export * from "./authenticationOptions";
+export * from "./botSwarmData";
+export * from "./connectionOptions";
+export * from "./plugin";
+export * from "./swarmBot";
